hdmitx21: Optimize frl training failure output strategy under uboot [1/1]
PD#SWPL-170535

Problem:
When the user selects frl mode for output, if frl training fails under
uboot,the output will be changed to 1080p. As a result, after entering
the kernel,the final output is 1080p instead of the frl mode selected
by the user. It is necessary to ensure that the final output under the
kernel is the frl mode selected by the user.

Solution:
When outputting frl mode, if frl training fails under uboot,
in order to ensure that it is displayed under uboot, change
to the default TMDS mode for output display. systemctrl
maintains the original 8k policy. After the subsequent systermctrl
starts running, if it is checked that the current output is not the
original frl mode, it will switch to the original frl mode.

Verify:
s7/s5

Change-Id: I3bc22f5dd6a5977eeec4427b8ab1ad7b6ae9fec9
Signed-off-by: xiang.wu1 <xiang.wu1@amlogic.com>