hdmitx: uboot modify default hdr_priority value [1/1]
PD#SWPL-191567

Problem:
The Google HDR strategy used by Android U is incompatible with
the private HDR strategy of Android R. When the default value
of hdr_priority is given as the Google HDR strategy value, it
will affect the private HDR strategy.

Solution:
For uboot 2023 currently used with androidu, a default value
of dv priority is given to hdr_priority. In other cases,
special modifications will be made for the project

Verify:
s7

Change-Id: Ic3055d8149dded322a4b5272ae5a9456707e7d6d
Signed-off-by: xiang.wu1 <xiang.wu1@amlogic.com>