ODROID-C5: board: Add fix DMA heap memory size for graphics
The default DMA heap memory size for the graphics is defined in the
device tree and it works with the resolution upto 1920x1080, but it
causes very bad display glitches when running Gnome Deskop in 4k
resolution. It requires more DMA pool for higher resolution.

One can set the DMA pool size by adding a line with the capacity to the
environment value with 'gfx-heap-size', or can define the same in
'config.ini'.

For example, for 180MB can be set

	[generic]
	...
	gfx-heap-size=180
	...

and its result is;

    $ od --endian=big -t x4 /proc/device-tree/reserved-memory/heap-gfx
    0000000 00000000 0b400000

This is identical to define the same amount in a device tree,

     dmaheap_cma_reserved:heap-gfx {
	     ...
	     size = <0x0 0xb400000>;
	     ...
     };

Change-Id: I7ba56385b97e654590a4a16059d8b4dd03eaa369